// backend-cluster/backend-v2/src/features/gitea/client/gitea-api.ts
export interface Activity {
  /** User represents a user */
  act_user?: User;
  /** @format int64 */
  act_user_id?: number;
  /** Comment represents a comment on a commit or issue */
  comment?: Comment;
  /** @format int64 */
  comment_id?: number;
  content?: string;
  /** @format date-time */
  created?: string;
  /** @format int64 */
  id?: number;
  is_private?: boolean;
  /** the type of action */
  op_type?:
    | "create_repo"
    | "rename_repo"
    | "star_repo"
    | "watch_repo"
    | "commit_repo"
    | "create_issue"
    | "create_pull_request"
    | "transfer_repo"
    | "push_tag"
    | "comment_issue"
    | "merge_pull_request"
    | "close_issue"
    | "reopen_issue"
    | "close_pull_request"
    | "reopen_pull_request"
    | "delete_tag"
    | "delete_branch"
    | "mirror_sync_push"
    | "mirror_sync_create"
    | "mirror_sync_delete"
    | "approve_pull_request"
    | "reject_pull_request"
    | "comment_pull"
    | "publish_release"
    | "pull_review_dismissed"
    | "pull_request_ready_for_review"
    | "auto_merge_pull_request";
  ref_name?: string;
  /** Repository represents a repository */
  repo?: Repository;
  /** @format int64 */
  repo_id?: number;
  /** @format int64 */
  user_id?: number;
}
// ...
export interface Attachment {
  browser_download_url?: string;
  /** @format date-time */
  created_at?: string;
  /** @format int64 */
  download_count?: number;
  /** @format int64 */
  id?: number;
  name?: string;
  /** @format int64 */
  size?: number;
  uuid?: string;
}
// ...
export interface Comment {
  assets?: Attachment[];
  body?: string;
  /** @format date-time */
  created_at?: string;
  html_url?: string;
  /** @format int64 */
  id?: number;
  issue_url?: string;
  original_author?: string;
  /** @format int64 */
  original_author_id?: number;
  pull_request_url?: string;
  /** @format date-time */
  updated_at?: string;
  /** User represents a user */
  user?: User;
}
// ...
export interface ExternalTracker {
  /** External Issue Tracker URL Format. Use the placeholders {user}, {repo} and {index} for the username, repository name and issue index. */
  external_tracker_format?: string;
  /** External Issue Tracker issue regular expression */
  external_tracker_regexp_pattern?: string;
  /** External Issue Tracker Number Format, either `numeric`, `alphanumeric`, or `regexp` */
  external_tracker_style?: string;
  /** URL of external issue tracker. */
  external_tracker_url?: string;
}

/** ExternalWiki represents setting for external wiki */
export interface ExternalWiki {
  /** URL of external wiki. */
  external_wiki_url?: string;
}
// ...
export interface InternalTracker {
  /** Let only contributors track time (Built-in issue tracker) */
  allow_only_contributors_to_track_time?: boolean;
  /** Enable dependencies for issues and pull requests (Built-in issue tracker) */
  enable_issue_dependencies?: boolean;
  /** Enable time tracking (Built-in issue tracker) */
  enable_time_tracker?: boolean;
}
// ...
export interface Permission {
  admin?: boolean;
  pull?: boolean;
  push?: boolean;
}
// ...
export interface RepoTransfer {
  /** User represents a user */
  doer?: User;
  /** User represents a user */
  recipient?: User;
  teams?: Team[];
}

/** Repository represents a repository */
export interface Repository {
  allow_fast_forward_only_merge?: boolean;
  allow_merge_commits?: boolean;
  allow_rebase?: boolean;
  allow_rebase_explicit?: boolean;
  allow_rebase_update?: boolean;
  allow_squash_merge?: boolean;
  archived?: boolean;
  /** @format date-time */
  archived_at?: string;
  avatar_url?: string;
  clone_url?: string;
  /** @format date-time */
  created_at?: string;
  default_allow_maintainer_edit?: boolean;
  default_branch?: string;
  default_delete_branch_after_merge?: boolean;
  default_merge_style?: string;
  description?: string;
  empty?: boolean;
  /** ExternalTracker represents settings for external tracker */
  external_tracker?: ExternalTracker;
  /** ExternalWiki represents setting for external wiki */
  external_wiki?: ExternalWiki;
  fork?: boolean;
  /** @format int64 */
  forks_count?: number;
  full_name?: string;
  has_actions?: boolean;
  has_issues?: boolean;
  has_packages?: boolean;
  has_projects?: boolean;
  has_pull_requests?: boolean;
  has_releases?: boolean;
  has_wiki?: boolean;
  html_url?: string;
  /** @format int64 */
  id?: number;
  ignore_whitespace_conflicts?: boolean;
  internal?: boolean;
  /** InternalTracker represents settings for internal tracker */
  internal_tracker?: InternalTracker;
  language?: string;
  languages_url?: string;
  licenses?: string[];
  link?: string;
  mirror?: boolean;
  mirror_interval?: string;
  /** @format date-time */
  mirror_updated?: string;
  name?: string;
  /** ObjectFormatName of the underlying git repository */
  object_format_name?: "sha1" | "sha256";
  /** @format int64 */
  open_issues_count?: number;
  /** @format int64 */
  open_pr_counter?: number;
  original_url?: string;
  /** User represents a user */
  owner?: User;
  /** Repository represents a repository */
  parent?: Repository;
  /** Permission represents a set of permissions */
  permissions?: Permission;
  private?: boolean;
  projects_mode?: string;
  /** @format int64 */
  release_counter?: number;
  /** RepoTransfer represents a pending repo transfer */
  repo_transfer?: RepoTransfer;
  /** @format int64 */
  size?: number;
  ssh_url?: string;
  /** @format int64 */
  stars_count?: number;
  template?: boolean;
  topics?: string[];
  /** @format date-time */
  updated_at?: string;
  url?: string;
  /** @format int64 */
  watchers_count?: number;
  website?: string;
}
// ...
export interface User {
  /** Is user active */
  active?: boolean;
  /** URL to the user's avatar */
  avatar_url?: string;
  /** @format date-time */
  created?: string;
  /** the user's description */
  description?: string;
  /** @format email */
  email?: string;
  /**
   * user counts
   * @format int64
   */
  followers_count?: number;
  /** @format int64 */
  following_count?: number;
  /** the user's full name */
  full_name?: string;
  /** URL to the user's gitea page */
  html_url?: string;
  /**
   * the user's id
   * @format int64
   */
  id?: number;
  /** Is the user an administrator */
  is_admin?: boolean;
  /** User locale */
  language?: string;
  /** @format date-time */
  last_login?: string;
  /** the user's location */
  location?: string;
  /** the user's username */
  login?: string;
  /**
   * the user's authentication sign-in name.
   * @default "empty"
   */
  login_name?: string;
  /** Is user login prohibited */
  prohibit_login?: boolean;
  /** Is user restricted */
  restricted?: boolean;
  /**
   * The ID of the user's Authentication Source
   * @format int64
   */
  source_id?: number;
  /** @format int64 */
  starred_repos_count?: number;
  /** User visibility level option: public, limited, private */
  visibility?: string;
  /** the user's website */
  website?: string;
}